refactor provider.js to Context
