+++ src/pages/index.js
@@ -4,7 +4,7 @@ import { makeStyles } from "@material-ui/core/styles";
 import CircularProgress from "@material-ui/core/CircularProgress";
 
 import Layout from "../components/layout/Layout";
-import { Context } from "../components/layout/Provider";
+import { Context } from "../components/layout/Context";
 import Splash from "../components/splash/Splash";
 import DirectoryPreview from "../components/splash/DirectoryPreview";
 import SEO from "../components/util/seo";
